requirements.txt
requests>=2.13.0
pytest>=3.1.3
vcrpy>=1.10.3
python-coveralls>=2.9.1
coverage>=4.0.3
nose>=1.3.7
